% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{.apply_footnotes}
\alias{.apply_footnotes}
\title{Attach targeted (row- or column-specific) footnotes to a gt table.}
\usage{
.apply_footnotes(gt_tbl, footnotes)
}
\description{
Attach targeted (row- or column-specific) footnotes to a gt table.
}
\keyword{internal}
