% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{.fmt_p}
\alias{.fmt_p}
\title{Format a p-value according to current clerk_options.}
\usage{
.fmt_p(
  x,
  p_digits = NULL,
  p_threshold = NULL,
  p_style = NULL,
  stars = NULL,
  star_thresholds = NULL
)
}
\description{
Format a p-value according to current clerk_options.
}
\keyword{internal}
