% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{.get_clerk_options}
\alias{.get_clerk_options}
\title{Returns current clerk options, always falling back to built-in defaults.}
\usage{
.get_clerk_options()
}
\description{
Returns current clerk options, always falling back to built-in defaults.
}
\keyword{internal}
