% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{.fmt_p_fdr}
\alias{.fmt_p_fdr}
\title{Format an FDR-corrected p-value, replacing non-survivors with ns label}
\usage{
.fmt_p_fdr(x, fdr_ns = NULL, fdr_threshold = NULL, fdr_ns_label = NULL, ...)
}
\description{
Format an FDR-corrected p-value, replacing non-survivors with ns label
}
\keyword{internal}
